+++ equ-clt/src/main/java/equ/clt/handler/bizerba/BizerbaHandler.java
@@ -66,9 +66,9 @@ public abstract class BizerbaHandler extends ScalesHandler {
     public void sendSoftCheck(SoftCheckInfo softCheckInfo) throws IOException {
     }
 
-    public Map<Long, SendTransactionBatch> sendTransaction(List<TransactionScalesInfo> transactionList, String charset, boolean encode) throws IOException {
+    public Map<Integer, SendTransactionBatch> sendTransaction(List<TransactionScalesInfo> transactionList, String charset, boolean encode) throws IOException {
 
-        Map<Long, SendTransactionBatch> sendTransactionBatchMap = new HashMap<>();
+        Map<Integer, SendTransactionBatch> sendTransactionBatchMap = new HashMap<>();
 
         Map<String, String> brokenPortsMap = new HashMap<>();
         if(transactionList.isEmpty()) {
